Revert "Promisify downloadWsdlToFolder"
This reverts commit fe3860e3e93100ceba4270861a110052ea141dd5.

+++ app/util.js
@@ -44,8 +44,9 @@ class Util {
     /**
      * Create WSDL for the provided path
      * @param {IResolveOptions} options
+     * @param {Function} cb
      */
-    async downloadWsdlToFolder(options) {
+    downloadWsdlToFolder(options, cb) {
 
         if (!options.apiVersion) {
             throw new Error("apiVersion is a required parameter");
@@ -61,20 +62,25 @@ class Util {
         const tempZip = `WSDL_v${options.apiVersion}_0.zip`;
         const remotePath = `http://www.netsuite.com/download/${tempZip}`;
 
-        try {
-            await fs.promises.access(folderPath, fs.constants.F_OK);
-        } catch (error) {
-            await fs.promises.mkdir(folderPath, { recursive: true });
-            
-            const fetch = new Promise((resolve, reject) => {
-                console.log('started');
+        fs.access(folderPath, fs.constants.F_OK, (fErr) => {
+            if (!fErr) {
+                cb(filePath);
+                return;
+            }
+
+            mkdirp(folderPath, (dErr) => {
+                if (dErr) {
+                    throw new Error(dErr.message);
+                }
+
                 agent
                     .get(remotePath)
                     .on("error", (gErr) => {
-                        reject(gErr.message);
+                        throw new Error(gErr.message);
                     })
                     .pipe(fs.createWriteStream(tempZip))
                     .on("finish", () => {
+
                         const zip = new AdmZip(tempZip);
                         zip.extractAllTo(folderPath, true);
                         fs.unlink(`./${tempZip}`, (uErr) => {
@@ -82,14 +88,11 @@ class Util {
                                 throw new Error(uErr.message);
                             }
                         });
-                        resolve(filePath);
+
+                        cb(filePath);
                     });
             });
-
-            return fetch.catch((error) => {
-                throw new Error(error);
-            });
-        }
+        });
     }
 
     /**
